moved runkit to createToken + feature toggle
