+++ src/components/containers/CreateToken.container.tsx
@@ -87,6 +87,7 @@ const createToken = (props: CreateTokenContainerProps): JSX.Element => (
     disabledButton={props.deleteApiToken.loading}
     network={props.network}
     textCreateTokenButton={getTextCreateTokenButton(props.network)}
+    email={props.user.profile.email}
   />
 )
