+++ src/config/features.ts
@@ -1,10 +1,15 @@
 export enum FeatureName {
   ToggleNetwork = 'toggle-network',
+  RunKit = 'run-kit',
 }
 
 export const initialFeatures: ReadonlyArray<any> = [
   {
     name: FeatureName.ToggleNetwork,
+    isActive: false,
+  },
+  {
+    name: FeatureName.RunKit,
     isActive: false,
   },
 ]
